requirements.txt
scrapy==2.4.1
requests==2.25.0
django==2.2
djangorestframework==3.11.0
pygments==2.7.3
django-cors-headers==3.6.0
python-decouple==3.3
psycopg2==2.8.6
drf-extensions==0.6.0
django-allauth==0.44.0
django-rest-authtoken==2.1.3
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework-jwt==1.11.0 